create a new file in util folder and separate a code of one time use remove duplicate repositories
